+++ integration/src/test/scala/org/knora/webapi/core/LayersTest.scala
@@ -47,6 +47,8 @@ import org.knora.webapi.slice.infrastructure.CacheManager
 import org.knora.webapi.slice.infrastructure.InfrastructureModule
 import org.knora.webapi.slice.infrastructure.api.ManagementEndpoints
 import org.knora.webapi.slice.infrastructure.api.ManagementRoutes
+import org.knora.webapi.slice.lists.api.ListsApiModule
+import org.knora.webapi.slice.lists.domain.ListsService
 import org.knora.webapi.slice.ontology.api.service.RestCardinalityService
 import org.knora.webapi.slice.ontology.api.service.RestCardinalityServiceLive
 import org.knora.webapi.slice.ontology.domain.service.CardinalityService
@@ -110,6 +112,7 @@ object LayersTest {
     InferenceOptimizationService &
     IriConverter &
     IriService &
+    ListsApiModule.Provided &
     ListsResponder &
     MessageRelay &
     OntologyCache &
@@ -146,7 +149,7 @@ object LayersTest {
     ApiV2Endpoints &
     HttpServer &
     IIIFRequestMessageHandler &
-    ListsResponderV2 &
+    ListsService &
     OntologyResponderV2 &
     RepositoryUpdater &
     ResourcesResponderV2 &
@@ -172,7 +175,9 @@ object LayersTest {
       IriConverter.layer,
       IriService.layer,
       KnoraResponseRenderer.layer,
+      ListsApiModule.layer,
       ListsResponder.layer,
+      ListsService.layer,
       ManagementEndpoints.layer,
       ManagementRoutes.layer,
       MessageRelayLive.layer,
@@ -217,7 +222,7 @@ object LayersTest {
       RepositoryUpdater.layer,
       HttpServer.layer,
       ApiV2Endpoints.layer,
-      ListsResponderV2.layer,
+      ListsService.layer,
       IIIFRequestMessageHandlerLive.layer,
       IriService.layer,
     )