+++ webapi/src/test/scala/org/knora/webapi/slice/admin/domain/model/GroupIriSpec.scala
@@ -5,16 +5,26 @@
 
 package org.knora.webapi.slice.admin.domain.model
 
+import zio.Chunk
 import zio.test.Gen
 import zio.test.Spec
 import zio.test.ZIOSpecDefault
 import zio.test.assertTrue
 import zio.test.check
 
+import org.knora.webapi.messages.OntologyConstants.KnoraAdmin.KnoraAdminPrefixExpansion
+
 object GroupIriSpec extends ZIOSpecDefault {
   override val spec: Spec[Any, Nothing] = suite("GroupIri should")(
     test("not be created from an empty value") {
       assertTrue(GroupIri.from("") == Left("Group IRI cannot be empty."))
+    },
+    test("allow prefixed builtin GroupIris") {
+      val builtIn = Chunk("UnknownUser", "KnownUser", "Creator", "ProjectMember", "ProjectAdmin", "SystemAdmin")
+        .map("knora-admin:" + _)
+      check(Gen.fromIterable(builtIn)) { it =>
+        assertTrue(GroupIri.from(it).map(_.value) == Right(it.replace("knora-admin:", KnoraAdminPrefixExpansion)))
+      }
     },
     test("be created from a valid value") {
       val validIris = Gen.fromIterable(
@@ -37,7 +47,7 @@ object GroupIriSpec extends ZIOSpecDefault {
           "http://rdfh.ch/groups/jDEEitJESRi3pDaDjjQ1WQ",
         ),
       )
-      check(invalidIris)(i => assertTrue(GroupIri.from(i) == Left(s"Group IRI is invalid.")))
+      check(invalidIris)(i => assertTrue(GroupIri.from(i) == Left(s"Group IRI is invalid: $i")))
     },
   )
 }